pyproject.toml
[project]
name = "travelhub"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.8",
    "django-jazzmin>=3.0.1",
    "django-mptt>=0.18.0",
    "net-tools>=0.1.2",
    "pillow>=12.0.0",
    "psycopg2-binary>=2.9.11",
    "requests>=2.32.5",
    "six>=1.17.0",
]

[dependency-groups]
dev = [
    "flake8>=7.3.0",
    "git-filter-repo>=2.47.0",
    "isort>=7.0.0",
    "python-dotenv>=1.2.1",
]
